Funcionalidades de HU 11 y HU 12

+++ src/main/java/com/kevin/emazon_cart/infraestructure/entity/CartEntity.java
@@ -24,8 +24,10 @@ public class CartEntity {
     private Long itemId;
 
     private Long userId;
-    private Double quantity;
+    private Long quantity;
     private Date updateDate;
     private Date creationDate;
+
+
 
 }
